Add Manifest piggy-backing optimization
Roo will now piggy-back manifest information onto outbound response
and delegated request messages. This eliminates the need for a
separate Manifest message in all cases except when a ServerTask
doesn't otherwise generate any outbound messages. The trade-off is
that all request and response headers are now larger, with enough
space to hold a single additional manifest structure. This design
was chosen because the cost of sending the extra bytes of header is
minimal compared to the overhead of sending and receiving a separate
Manifest message.