Redesign ping/pong protocol to reduce traffic
Change the ping/pong protocol so that each client ping message only
generates a single pong message back to the client. This reduces
the incast of pong messages back to the client when timeouts occur.
Additionally, changes are made to skip ping processing on servers
when the server is only contacted to forward pings as a proxy; this
reduces unnecessary ping traffic.